+++ overworld/src/components/Graveyard/BaseHero copy.js
@@ -11,9 +11,15 @@ const baseHero = {
   heroY: globalVars.heroCenterY,
   targetHeroX: globalVars.heroCenterX,
   targetHeroY: globalVars.heroCenterY,
-  dashBoost: 2,
-  baseVel: .7,
-  currentVel: .7,
+  topDashBoost: 1,
+  dashBoost: 0,
+  boostMaxVel: 6,
+  baseMaxVel: 4,
+  maxVel: 4,
+  rateAccel: 1,
+  rateDecel: 1,
+  xVel: 0,
+  yVel: 0,
   baseMoveSpeed: 20,
   moveSpeed: 20,
   frameCountLimiter: 0,
